gcp/requirements.txt
functions-framework==3.*
tensorflow>=2.18.0
keras>=3.0.0
google-cloud-storage
Pillow
flask